Start applying pymongo's tests to async wrapper, which I'm going to start calling "motor" in these comments. Major bugfixing and new features ensue.

+++ pymongo/tornado_async/test_tornado_async/eventually.py
@@ -20,7 +20,7 @@ class AssertEventuallyTest(unittest.TestCase):
     ):
         if timeout_sec is None:
             timeout_sec = 5
-        timeout_sec = max(timeout_sec, int(os.environ.get('TIMEOUT_SEC', 0)))
+        timeout_sec = max(timeout_sec, float(os.environ.get('TIMEOUT_SEC', 0)))
         start = time.time()
         loop = ioloop.IOLoop.instance()
